# openapi.yml
openapi: '3.1.0'
info:
  title: Comprehensive Books Management REST API
  description: >-
    **A REST API for managing Users, Books, Authors, Genres, and their interactions, developed with [Express](https://expressjs.com).**

    This API provides endpoints to manage Users, Books, Authors, Genres, and the relationships between them.
  version: 1.0.0
  contact:
    url: https://github.com/Broannn/BookNest-RestAPI
servers:
  - url: https://booknest-restapi.onrender.com/
...
paths:
...
  /  /books:
    post:
      summary: Create a book.
      tags:
        - books
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookData'
      responses:
        201:
          description: The book was successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
    get:
      summary: List books.
      tags:
        - books
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        200:
          description: A list of books.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
  /books/{bookId}:
    get:
      summary: Get a book by ID.
      tags:
        - books
      parameters:
        - name: bookId
          in: path
          required: true
          description: The ID of the book.
          schema:
            type: string
      responses:
        200:
          description: The book details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        404:
          description: Book not found.
    put:
      summary: Update a book.
      tags:
        - books
      parameters:
        - name: bookId
          in: path
          required: true
          description: The ID of the book.
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookData'
      responses:
        200:
          description: Book successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        404:
          description: Book not found.
    delete:
      summary: Delete a book.
      tags:
        - books
      parameters:
        - name: bookId
          in: path
          required: true
          description: The ID of the book.
          schema:
            type: string
      responses:
        204:
          description: Book successfully deleted.
        404:
          description: Book not found.
  /books/{bookId}/genres:
    post:
      summary: Add a genre to a book.
      tags:
        - genres
      parameters:
        - name: bookId
          in: path
          required: true
          description: The ID of the book.
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                genreId:
                  type: string
                  description: The ID of the genre to add.
      responses:
        201:
          description: Genre added to the book.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookGenre'
    get:
      summary: Get genres of a book.
      tags:
        - genres
      parameters:
        - name: bookId
          in: path
          required: true
          description: The ID of the book.
          schema:
            type: string
      responses:
        200:
          description: A list of genres for the book.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Genre'
    delete:
      summary: Remove a genre from a book.
      tags:
        - genres
      parameters:
        - name: bookId
          in: path
          required: true
          description: The ID of the book.
          schema:
            type: string
        - name: genreId
          in: path
          required: true
          description: The ID of the genre to remove.
          schema:
            type: string
      responses:
        204:
          description: Genre successfully removed from the book.
  /books/{bookId}/critiques:
    get:
      summary: Get all critiques for a book.
      tags:
        - critiques
      parameters:
        - name: bookId
          in: path
          required: true
          description: The ID of the book.
          schema:
            type: string
      responses:
        200:
          description: A list of critiques for the book.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Critique'
    post:
      summary: Add a critique to a book.
      tags:
        - critiques
      parameters:
        - name: bookId
          in: path
          required: true
          description: The ID of the book.
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  description: The ID of the user adding the critique.
                rating:
                  type: number
                  description: The rating given to the book.
                comment:
                  type: string
                  description: The critique comment.
      responses:
        201:
          description: Critique successfully added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Critique'
    delete: # Correct indentation here
      summary: Delete a critique for a book.
      tags:
        - critiques
      parameters:
        - name: bookId
          in: path
          required: true
          description: The ID of the book.
          schema:
            type: string
        - name: critiqueId
          in: path
          required: true
          description: The ID of the critique to be deleted.
          schema:
            type: string
      responses:
        204:
          description: Critique successfully deleted.
        404:
          description: Critique not found.
...
components:
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
      example: 1
    pageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
      example: 25
...
  schemas:
    Genre: # Fixed indentation for Genre schema
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the genre.
          example: 5fbd827a512e73129cfedb1c
        name:
          type: string
          description: The name of the genre.
          example: Fiction
        createdAt:
          type: string
          format: date-time
          description: The date the genre was created.
          example: 2025-01-13T10:00:00.000Z
      required:
        - id
        - name
...
    Book: # Corrected indentation starts here
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the book.
          example: 123bookid
        title:
          type: string
          description: The title of the book.
          example: Great Expectations
        authorId:
          type: string
          description: The ID of the author of the book.
          example: 7a9bc839033eedf0f
        genres:
          type: array
          description: Genres associated with the book.
          items:
            $ref: '#/components/schemas/Genre'
        publicationDate:
          type: string
          format: date
          description: The publication date of the book.
          example: 2023-01-01
      required:
        - id
        - title
        - authorId
    BookData:
      type: object
      properties:
        title:
          type: string
          description: The title of the book.
          example: Great Expectations
        authorId:
          type: string
          description: The ID of the author of the book.
          example: 7a9bc839033eedf0f
        genres:
          type: array
          description: Genres to associate with the book.
          items:
            $ref: '#/components/schemas/Genre'
        publicationDate:
          type: string
          format: date
          description: The publication date of the book.
          example: 2023-01-01
      required:
        - title
        - authorId
    Critique:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the critique.
          example: 456critiqueid
        bookId:
          type: string
          description: The ID of the critiqued book.
          example: 123bookid
        userId:
          type: string
          description: The ID of the user who wrote the critique.
          example: 7fe91a9a41bc839033eedf0f
        rating:
          type:
            type: number
            description: The rating given to the book.
            example: 4.5